address PR feedback: 3 files changed
